Add new base classes for queries
Introduce SingleObjectQuery and ListQuery as new base classes for our
queries (more explicit for fields that act as query).

With using this new base classes the queries can be implemented in a more
declarative way instead of heaving to always override the constructor.
See the also included examples for GetFeed and GetFeeds.