Refactor auto deletion of audit/task reports
Drop auto delete data field and only use auto delete that expects a
number. It is not necessary to have two fields for setting an auto
deletion of reports. If auto delete is not set all reports are kept and
no report is deleted automatically. If auto delete is set to a number
only the latest number of reports are kept for the audit or task.